+++ unit_tests/test_fuzz_logger_csv.py
@@ -28,7 +28,7 @@ class TestFuzzLoggerCsvFreeFunctions(unittest.TestCase):
         Then: Hex of several-line string is output first, then repr format.
         """
         given = b"abc\n123\r\nA\n"
-        expected = u"61 62 63 0a 31 32 33 0d 0a 41 0a"
+        expected = "61 62 63 0a 31 32 33 0d 0a 41 0a"
         self.assertEqual(expected, fuzz_logger_csv.hex_to_hexstr(given))