+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Upcoming
+--------
+Features
+^^^^^^^^
+- Added support for fuzzing NETCONF servers with the `NETCONFConnection` class.
+
 v0.4.0
 ------
 Features
@@ -14,7 +20,6 @@ Features
 - Added `Simple` primitive that uses only the specified values for fuzzing.
 - Added `Float` primitive with support for IEEE 754 encoding.
 - Added an example for s_float/Float usage.
-- Added support for fuzzing NETCONF servers with the `NETCONFConnection` class.
 
 Fixes
 ^^^^^