+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Upcoming
 Features
 ^^^^^^^^
 - Added support for fuzzing NETCONF servers with the `NETCONFConnection` class.
+- Add support and tests for Python 3.10
 
 Fixes
 ^^^^^